Upgrading the carbon kernel version to 5.2.0-m4 and refactoring changes due to upgrade
