+++ components/data-bridge/org.wso2.carbon.databridge.agent/src/main/java/org/wso2/carbon/databridge/agent/AgentHolder.java
@@ -19,6 +19,7 @@ package org.wso2.carbon.databridge.agent;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.wso2.carbon.config.provider.ConfigProvider;
 import org.wso2.carbon.databridge.agent.conf.Agent;
 import org.wso2.carbon.databridge.agent.conf.AgentConfiguration;
 import org.wso2.carbon.databridge.agent.conf.DataAgentConfigurationFileResolver;
@@ -27,7 +28,6 @@ import org.wso2.carbon.databridge.agent.exception.DataEndpointAgentConfiguration
 import org.wso2.carbon.databridge.agent.exception.DataEndpointException;
 import org.wso2.carbon.databridge.agent.internal.DataAgentServiceValueHolder;
 import org.wso2.carbon.databridge.agent.util.DataEndpointConstants;
-import org.wso2.carbon.kernel.configprovider.ConfigProvider;
 import org.yaml.snakeyaml.Yaml;
 
 import java.io.File;
@@ -125,7 +125,7 @@ public class AgentHolder {
                     dataAgentsConfiguration = DataAgentConfigurationFileResolver.
                             resolveAndSetDataAgentConfiguration
                                     ((LinkedHashMap) configProvider.
-                                            getConfigurationMap(DataEndpointConstants.DATA_AGENT_CONFIG_NAMESPACE));
+                                            getConfigurationObject(DataEndpointConstants.DATA_AGENT_CONFIG_NAMESPACE));
                 }
             } else {
                 File file = new File(configPath);
