+++ components/data-bridge/org.wso2.carbon.databridge.core/pom.xml
@@ -89,13 +89,10 @@
             <groupId>org.apache.log4j.wso2</groupId>
             <artifactId>log4j</artifactId>
         </dependency>
-
-
-        <!--<dependency>-->
-        <!--<groupId>org.wso2.securevault</groupId>-->
-        <!--<artifactId>org.wso2.securevault</artifactId>-->
-        <!--</dependency>-->
-
+        <dependency>
+            <groupId>org.wso2.carbon.config</groupId>
+            <artifactId>org.wso2.carbon.config</artifactId>
+        </dependency>
     </dependencies>
     <build>
 
@@ -105,9 +102,9 @@
                 <artifactId>maven-surefire-plugin</artifactId>
             </plugin>
             <plugin>
-                <groupId>org.wso2.carbon</groupId>
-                <artifactId>org.wso2.carbon.extensions.configuration.maven.plugin</artifactId>
-                <version>${carbon.kernel.version}</version>
+                <groupId>org.wso2.carbon.config</groupId>
+                <artifactId>org.wso2.carbon.config.maven.plugin</artifactId>
+                <version>${carbon.config.version}</version>
                 <executions>
                     <execution>
                         <goals>
