+++ components/data-bridge/org.wso2.carbon.databridge.core/src/main/java/org/wso2/carbon/databridge/core/conf/DataBridgeConfiguration.java
@@ -15,8 +15,9 @@
  */
 package org.wso2.carbon.databridge.core.conf;
 
-import org.wso2.carbon.kernel.annotations.Configuration;
-import org.wso2.carbon.kernel.annotations.Element;
+
+import org.wso2.carbon.config.annotation.Configuration;
+import org.wso2.carbon.config.annotation.Element;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
