+++ pom.xml
@@ -404,6 +404,14 @@
                 <type>zip</type>
             </dependency>
 
+
+            <!-- Latest Dependencies - after 5.2.0-m4 migration -->
+            <dependency>
+                <groupId>org.wso2.carbon.config</groupId>
+                <artifactId>org.wso2.carbon.config</artifactId>
+                <version>${carbon.config.version}</version>
+            </dependency>
+
         </dependencies>
     </dependencyManagement>
 
@@ -463,9 +471,9 @@
 
         <osgi.framework.import.version.range>[1.8.0, 2.0.0)</osgi.framework.import.version.range>
         <osgi.service.tracker.import.version.range>[1.5.1, 2.0.0)</osgi.service.tracker.import.version.range>
-        <carbon.kernel.version>5.2.0-m3</carbon.kernel.version>
+        <carbon.kernel.version>5.2.0-m4</carbon.kernel.version>
         <carbon.kernel.package.import.version.range>[5.0.0, 6.0.0)</carbon.kernel.package.import.version.range>
-        <carbon.kernel.pax.version>5.2.0-m3</carbon.kernel.pax.version>
+        <carbon.kernel.pax.version>5.2.0-m4</carbon.kernel.pax.version>
         <slf4j.version>1.7.5</slf4j.version>
         <gson.version>2.8.0</gson.version>
         <disruptor.version>3.3.2.wso2v2</disruptor.version>
@@ -481,6 +489,9 @@
         <disruptor.version.range>[2.10.0,3.4)</disruptor.version.range>
         <carbon.security.caas.version>1.0.2-m1</carbon.security.caas.version>
         <org.wso2.json.version>3.0.0.wso2v1</org.wso2.json.version>
+
+        <!-- Latest Dependencies -->
+        <carbon.config.version>2.0.0</carbon.config.version>
 
     </properties>
